requirements.txt
asgiref==3.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
django-debug-toolbar==3.2.4
django-filter==21.1
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
drf-yasg==1.21.4
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
Pillow==9.1.0
psycopg2==2.9.9
pytz==2022.1
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sqlparse==0.4.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==2.2.1
django_ckeditor==6.7.1